Refactor project structure: rename and reorganize files for improved clarity and maintainability
