+++ README.md
@@ -1,0 +1,3 @@
+# WEMbed
+
+A Python Context Engineering Library for LLMs
