+++ requirements.txt
@@ -1,0 +1,114 @@
+﻿accelerate==1.10.1
+annotated-types==0.7.0
+anyio==4.10.0
+attrs==25.3.0
+beautifulsoup4==4.13.5
+certifi==2025.8.3
+charset-normalizer==3.4.3
+click==8.3.0
+click-default-group==1.2.4
+colorama==0.4.6
+condense-json==0.1.3
+dill==0.4.0
+distro==1.9.0
+docling==2.53.0
+docling-core==2.48.1
+docling-ibm-models==3.9.1
+docling-parse==4.5.0
+easyocr==1.7.2
+et_xmlfile==2.0.0
+Faker==37.8.0
+filelock==3.19.1
+filetype==1.2.0
+fsspec==2025.9.0
+greenlet==3.2.4
+h11==0.16.0
+httpcore==1.0.9
+httpx==0.28.1
+huggingface-hub==0.35.0
+idna==3.10
+imageio==2.37.0
+Jinja2==3.1.6
+jiter==0.11.0
+jsonlines==3.1.0
+jsonref==1.1.0
+jsonschema==4.25.1
+jsonschema-specifications==2025.9.1
+latex2mathml==3.78.1
+lazy_loader==0.4
+llm==0.27.1
+llm-ollama==0.14.0
+lxml==5.4.0
+markdown-it-py==4.0.0
+marko==2.2.0
+MarkupSafe==3.0.2
+mdurl==0.1.2
+mpire==2.10.2
+mpmath==1.3.0
+multiprocess==0.70.18
+networkx==3.5
+ninja==1.13.0
+numpy==2.3.3
+ollama==0.5.4
+openai==1.108.0
+opencv-python-headless==4.11.0.86
+openpyxl==3.1.5
+packaging==25.0
+pandas==2.3.2
+pillow==11.3.0
+pluggy==1.6.0
+polyfactory==2.22.2
+psutil==7.1.0
+psycopg2-binary==2.9.10
+puremagic==1.30
+pyclipper==1.3.0.post6
+pydantic==2.11.9
+pydantic-settings==2.10.1
+pydantic_core==2.33.2
+Pygments==2.19.2
+pylatexenc==2.10
+pypdfium2==4.30.0
+pyreadline3==3.5.4
+python-bidi==0.6.6
+python-dateutil==2.9.0.post0
+python-docx==1.2.0
+python-dotenv==1.1.1
+python-pptx==1.0.2
+python-ulid==3.1.0
+pytz==2025.2
+pywin32==311
+PyYAML==6.0.2
+referencing==0.36.2
+regex==2025.9.18
+requests==2.32.5
+rich==14.1.0
+rpds-py==0.27.1
+rtree==1.4.1
+safetensors==0.6.2
+scikit-image==0.25.2
+scipy==1.16.2
+semchunk==2.2.2
+setuptools==80.9.0
+shapely==2.1.1
+shellingham==1.5.4
+six==1.17.0
+sniffio==1.3.1
+soupsieve==2.8
+SQLAlchemy==2.0.43
+sqlite-fts4==1.0.3
+sqlite-migrate==0.1b0
+sqlite-utils==3.38
+sympy==1.14.0
+tabulate==0.9.0
+tifffile==2025.9.9
+tokenizers==0.22.0
+torch==2.8.0
+torchvision==0.23.0
+tqdm==4.67.1
+transformers==4.56.1
+typer==0.16.1
+typing-inspection==0.4.1
+typing_extensions==4.15.0
+tzdata==2025.2
+urllib3==2.5.0
+xlsxwriter==3.2.9
